refactor: remove use of ToolchainObject
not needed with the latest version of `client.Object`
